+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 All notable changes to the "amiga-debug" extension will be documented in this file.
 
 ## 1.5.4
-- NEW: profiler: improve blitter tooltip for line mode, show names of commonly used minterms
+- NEW: profiler: improve blitter tooltip for line mode, show simplified formulas of minterms
 - NEW: profiler: show blitter mode (`Blit`, `Line`, `Fill`, `Clear`) in blitter timeline
 
 ## 1.5.3